fix menu placement
