+++ components/misc/navbarMenu.vue
@@ -1,6 +1,6 @@
 <template>
   <v-list>
-    <div v-if="logged">
+    <v-list-item-group v-if="logged">
       <v-list-item>
         <v-list-item-avatar>
           <v-avatar size="36px" :color="randomColor">
@@ -17,11 +17,22 @@
           <v-list-item-subtitle>{{ userData.email }}</v-list-item-subtitle>
         </v-list-item-content>
       </v-list-item>
-    </div>
-    <v-list-item>
-      <misc-theme-selector />
-    </v-list-item>
+    </v-list-item-group>
+
     <v-list-item-group>
+      <v-list-item>
+        <v-list-item-icon>
+          <v-icon>mdi-black-mesa</v-icon>
+        </v-list-item-icon>
+
+        <v-list-item-content>
+          <v-list-item-title>Black hole mode</v-list-item-title>
+        </v-list-item-content>
+
+        <v-list-item-action>
+          <v-switch v-model="isDark"></v-switch>
+        </v-list-item-action>
+      </v-list-item>
       <v-list-item @click="$emit('loginClick')">
         <v-list-item-icon>
           <v-icon>mdi-account</v-icon>
@@ -43,5 +54,13 @@ export default class NavbarMenu extends Vue {
   @Prop(Object) userData
   @Prop(String) loginText
   @Prop(String) randomColor
+
+  get isDark() {
+    return this.$vuetify.theme.isDark
+  }
+
+  set isDark(val) {
+    this.$vuetify.theme.isDark = val
+  }
 }
 </script>
